Make source coverage view compact
